+++ halotools/mock_observables/two_point_clustering/tests/test_tpcf_jackknife.py
@@ -1,31 +1,35 @@
-#!/usr/bin/env python
+""" Module providing unit-testing for the `~halotools.mock_observables.tpcf_jackknife` function. 
+"""
+from __future__ import absolute_import, division, print_function, unicode_literals
 
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
 import numpy as np
-import pytest 
+from astropy.tests.helper import pytest
+from astropy.utils.misc import NumpyRNGContext
 
 from ..tpcf_jackknife import tpcf_jackknife
-from ..tpcf import tpcf
+from ...tpcf import tpcf
 
 slow = pytest.mark.slow
 
 __all__=['test_tpcf_jackknife_corr_func', 'test_tpcf_jackknife_cov_matrix']
 
 #create toy data to test functions
-Npts=100
-sample1 = np.random.random((Npts,3))
-randoms = np.random.random((Npts*10,3))
 period = np.array([1.0,1.0,1.0])
 rbins = np.linspace(0.001,0.3,5).astype(float)
 rmax = rbins.max()
+
+fixed_seed = 43 
 
 @pytest.mark.slow
 def test_tpcf_jackknife_corr_func():
     """
     test the correlation function
     """
-    
+    Npts=100
+    with NumpyRNGContext(fixed_seed):
+        sample1 = np.random.random((Npts,3))
+        randoms = np.random.random((Npts*10,3))
+        
     result_1,err = tpcf_jackknife(sample1, randoms, rbins, 
         Nsub=5, period = period, num_threads=1)
 
@@ -39,6 +43,10 @@ def test_tpcf_jackknife_cov_matrix():
     """
     test the covariance matrix
     """
+    Npts=100
+    with NumpyRNGContext(fixed_seed):
+        sample1 = np.random.random((Npts,3))
+        randoms = np.random.random((Npts*10,3))
     
     nbins = len(rbins)-1
